Split analysis job sample metadata helpers
